Updated emu directory with files from penguin

+++ emu/singularity/emu_adj.sh
@@ -1,4 +1,4 @@
-#!/bin/bash 
+#!/bin/bash -e
 
 #=================================
 # Shell script for V4r4 Adjoint Tool
@@ -8,9 +8,8 @@
 #    3) do_adj.csh
 #=================================
 
-native_singularity=NATIVE_SINGULARITY
 singularity_image=SINGULARITY_IMAGE
-native_setup=NATIVE_SETUP
+emu_input_dir=EMU_INPUT_DIR
 
 echo " "
 echo "************************************"
@@ -19,24 +18,47 @@ echo "************************************"
 
 echo ${PWD} > emu.fcwd 
 
-/bin/rm -f my_commands.sh
-echo '#!/bin/bash' > my_commands.sh & chmod +x my_commands.sh
-echo 'cd /inside_out'               >> my_commands.sh
+# Step 0: Check required EMU Input 
+fdum=${emu_input_dir}/forcing/other/flux-forced/forcing
+if [[ ! -d $fdum ]]; then 
+    echo " "
+    echo "ABORT: EMU Input for Adjoint Tool not found;"
+    echo $fdum
+    echo "Run PUBLICDIR/emu_download_input.sh"
+    echo "using ${emu_input_dir} as 'directory name to place EMU Input'" 
+    echo "to download forcing needed for the Adjoint Tool." 
+    exit 1
+fi
 
 # ------------------------------------------
 # Step 1: Tool Setup
-echo 'echo " "'                           >> my_commands.sh
-echo 'echo "**** Step 1: Tool Setup"'     >> my_commands.sh
-echo 'echo "     Running setup_adj.csh "' >> my_commands.sh
-echo '${basedir}/emu/setup_adj.sh'        >> my_commands.sh
+echo " "
+echo "**** Step 1: Tool Setup"
+echo "     Running setup_adj.csh "
+
+# Initialize my_commands.sh
+/bin/rm -f my_commands.sh
+echo '#!/bin/bash -e' > my_commands.sh && chmod +x my_commands.sh
+echo 'cd /inside_out'               >> my_commands.sh
+echo '${emu_dir}/emu/setup_adj.sh'        >> my_commands.sh
+
+singularity exec --bind ${PWD}:/inside_out \
+     ${singularity_image} /inside_out/my_commands.sh
 
 # Step 2: Specification
-echo 'echo " "'                          >> my_commands.sh
-echo 'echo "**** Step 2: Specification"' >> my_commands.sh
-echo 'echo "     Running adj.x"'         >> my_commands.sh
+echo " "
+echo "**** Step 2: Specification"
+echo "     Running adj.x"
+
+# Initialize my_commands.sh
+/bin/rm -f my_commands.sh
+echo '#!/bin/bash -e' > my_commands.sh && chmod +x my_commands.sh
+echo 'cd /inside_out'               >> my_commands.sh
+cp -f PUBLICDIR/mitgcm_timing.nml .
+echo /emu_input_dir > ./input_setup_dir
 echo './adj.x'                           >> my_commands.sh
 
-${native_singularity} exec --bind ${PWD}:/inside_out \
+singularity exec --bind ${emu_input_dir}:/emu_input_dir:ro --bind ${PWD}:/inside_out \
      ${singularity_image} /inside_out/my_commands.sh
 
 if [ -f "adj.dir_out" ] && [ -f "pbs_adj.sh" ]; then
@@ -54,18 +76,18 @@ echo "**** Step 3: Calculation"
 
 returndir=$PWD
 
+cd ${rundir}
+
 # prepare script for 1) inside Singularity 
 echo "  1) Set up files for MITgcm "
 
-cd ${rundir}
-
 /bin/rm -f my_commands.sh 
-echo '#!/bin/bash'     > my_commands.sh & chmod +x my_commands.sh 
+echo '#!/bin/bash -e' > my_commands.sh && chmod +x my_commands.sh
 echo 'cd /inside_out'                   >> my_commands.sh
-echo '${basedir}/emu/singularity/setup_forcing.sh' >> my_commands.sh
-echo '${basedir}/emu/singularity/do_adj_prep.sh' >> my_commands.sh
+echo '${emu_dir}/emu/singularity/setup_forcing.sh' >> my_commands.sh
+echo '${emu_dir}/emu/singularity/do_adj_prep.sh' >> my_commands.sh
 
-${native_singularity} exec --bind ${native_setup}:/emu_outside:ro --bind ${rundir}:/inside_out \
+singularity exec --bind ${emu_input_dir}:/emu_input_dir:ro --bind ${rundir}:/inside_out \
      ${singularity_image} /inside_out/my_commands.sh
 
 cd ${returndir}
@@ -73,9 +95,9 @@ cd ${returndir}
 # submit batch job to do 2)
 echo "  2) Run MITgcm adjoint "
 
-BATCH_COMMAND pbs_adj.sh
+#BATCH_COMMAND pbs_adj.sh
 
-echo "... Batch job pbs_adj.sh has been submitted "
+echo "... Running batch job pbs_adj.sh "
 echo "    to compute the model's adjoint gradients." 
 
 echo " "
@@ -89,3 +111,12 @@ echo "    Results will be in " ${dum}
 echo '********************************************'
 echo " "
 
+dum="${dum%output}temp"
+echo "Progress of the computation can be monitored by"
+echo "  grep ad_time_tsnumber ${dum}/STDOUT.0000 "
+echo "which lists the model's time-step (one-hour) at 10-day"
+echo "intervals backward from the target instant to the model's"
+echo "initial time (hour 0), 01 January 1992 12Z." 
+echo " "
+
+BATCH_COMMAND pbs_adj.sh
